+++ Backend/Data/ApplicationDbContext.cs
@@ -1,4 +1,5 @@
 using Microsoft.EntityFrameworkCore;
+using Backend.Models.Entities;
 
 namespace Backend.Data
 {
@@ -7,5 +8,14 @@ namespace Backend.Data
         public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : base(options)
         {
         }
+
+      
+      public DbSet<WordType> WordTypes { get; set; }
+      public DbSet<Word> Words { get; set; }
+      public DbSet<User> Users { get; set; }    
+
+      public DbSet<Sentence> Sentences { get; set; }
+    public DbSet<SentenceWord> SentenceWords { get; set; }
+
     }
 }