+++ Backend/Models/Entities/WordType.cs
@@ -1,0 +1,14 @@
+
+
+namespace Backend.Models.Entities
+{
+    public class WordType
+    {
+      public int Id { get; set; }
+
+    public string Name { get; set; } = string.Empty;
+
+    // Navigation Property
+    public ICollection<Word> Words { get; set; } = new List<Word>();
+    }
+}